Add Homepage image slider(SlickSlider)
