+++ Client/package.json
@@ -26,6 +26,7 @@
     "react-scripts": "5.0.1",
     "react-slick": "^0.29.0",
     "react-transition-group": "^4.4.5",
+    "slick-carousel": "^1.8.1",
     "swiper": "^8.4.4",
     "typescript": "^4.8.2",
     "web-vitals": "^2.1.4"
