+++ Client/src/component/components.css
@@ -728,6 +728,105 @@ ul {
   display: none;
 }
 
+
+
+.single-img h2 {
+  position:absolute;
+  bottom: 5%;
+  z-index: 1;
+  left: 8%;
+  color: lawngreen;
+  font-family: Cardo;
+  font-size: 35px;
+}
+
+.img-area {
+  position: relative;
+  z-index: 1;
+  height: 100%;
+  max-width: 100%;
+  margin: 50px auto;
+}
+
+
+
+
+.img-slide-track{
+  transition: all 1s cubic-bezier(0.7, 0, 0.3, 1);
+}
+
+ .single-img{
+  height:700px;
+  position: relative;
+  z-index: 1;
+}
+
+ .single-img img{
+  width: 100%;
+  height: 100%;
+  transition: all 1s cubic-bezier(0.7, 0, 0.3, 1);
+  transform: scale(1.2);
+}
+
+ .single-img.slick-active img{
+  -webkit-animation: kenburns-top-right 8s ease-out both;
+        animation: kenburns-top-right 8s ease-out both;
+}
+
+
+
+.kenburns-top-right {
+-webkit-animation: kenburns-top-right 5s ease-out both;
+        animation: kenburns-top-right 5s ease-out both;
+}
+
+/* ----------------------------------------------
+* Generated by Animista on 2022-10-19 12:4:11
+* Licensed under FreeBSD License.
+* See http://animista.net/license for more info. 
+* w: http://animista.net, t: @cssanimista
+* ---------------------------------------------- */
+
+/**
+* ----------------------------------------
+* animation kenburns-top-right
+* ----------------------------------------
+*/
+@-webkit-keyframes kenburns-top-right {
+  0% {
+    -webkit-transform: scale(1) translate(0, 0);
+            transform: scale(1) translate(0, 0);
+    -webkit-transform-origin: 84% 16%;
+            transform-origin: 84% 16%;
+  }
+  100% {
+    -webkit-transform: scale(1.25) translate(20px, -15px);
+            transform: scale(1.25) translate(20px, -15px);
+    -webkit-transform-origin: right top;
+            transform-origin: right top;
+  }
+}
+@keyframes kenburns-top-right {
+  0% {
+    -webkit-transform: scale(1) translate(0, 0);
+            transform: scale(1) translate(0, 0);
+    -webkit-transform-origin: 84% 16%;
+            transform-origin: 84% 16%;
+  }
+  100% {
+    -webkit-transform: scale(1.25) translate(20px, -15px);
+            transform: scale(1.25) translate(20px, -15px);
+    -webkit-transform-origin: right top;
+            transform-origin: right top;
+  }
+}
+
+.slick-current{ 
+  -webkit-animation: kenburns-top-right 8s ease-out both;
+        animation: kenburns-top-right 8s ease-out both;
+}
+
+
 @media only screen and (max-width: 1220px) {
   .ContactImg {
     display: none;
@@ -786,6 +885,10 @@ ul {
     margin-bottom: 25px;
     text-decoration: none;
   }
+
+  
+  
+
 }
 
 @media only screen and (max-width: 600px) {
